+++ common.h
@@ -257,7 +257,7 @@ int get_wwnid_from_pretty
 int check_dir
 	(char *);
 void check_overflow
-	(unsigned long long, unsigned long long, unsigned long long);
+	(unsigned int, unsigned int, unsigned int);
 
 #ifndef SOURCE_SADC
 int count_bits
